Effect of companion - Health and Damage

+++ HH3/HH3.pde
@@ -52,7 +52,7 @@ void setup() {
 }
 float healthmax = 3;
 float health = healthmax;
-int healthcounter = 0;
+float healthcounter = 0;
 //if healthcounter is 0 and you collide, then health -=.5 and health counter = 1;
 
 long tick;
